+++ js/data/const.js
@@ -1,0 +1,3 @@
+export const NAMES = [
+  'Вася', 'Маша', 'Даша', 'Глаша', 'Саша', 'Мася', 'Дуся', 'Муся',
+];
